requirements.txt
fastapi
uvicorn
python-telegram-bot
python-dotenv
httpx
requests
python-multipart
pytesseract
Pillow
openai-whisper
